requirements.txt
beautifulsoup4==4.13.4
gspread==6.2.1
gspread_dataframe==4.0.0
numpy==2.2.5
pandas==2.2.3
protobuf==6.31.0
psycopg2_binary==2.9.10
Requests==2.32.3
SQLAlchemy==2.0.41
